+++ src/Components/Components/test/CascadingModelBinderTest.cs
@@ -198,7 +198,6 @@ public class CascadingModelBinderTest
         {
             builder.OpenComponent<CascadingModelBinder>(0);
             builder.AddAttribute(1, nameof(CascadingModelBinder.Name), contextName);
-            builder.AddAttribute(2, nameof(CascadingModelBinder.IsFixed), true);
             builder.AddAttribute(3, nameof(CascadingModelBinder.ChildContent), contents);
             builder.CloseComponent();
         });
@@ -222,7 +221,6 @@ public class CascadingModelBinderTest
         var testComponent = new TestComponent(builder =>
         {
             builder.OpenComponent<CascadingModelBinder>(0);
-            builder.AddAttribute(1, nameof(CascadingModelBinder.IsFixed), isFixed);
             builder.AddAttribute(2, nameof(CascadingModelBinder.ChildContent), contents);
             builder.CloseComponent();
         });
